requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-filter==21.1
django-redis==4.12.1
djangorestframework==3.15.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
Markdown==3.3.7
packaging==21.3
pycountry==22.3.5
pyparsing==3.0.7
pytz==2024.2
redis==3.5.3
sqlparse==0.4.4
typing_extensions==4.1.1
zipp==3.6.0
